feat: connecting lvd

+++ src/Tubes3_apaKek/Models/ResultData.cs
@@ -2,7 +2,7 @@ namespace Tubes3_apaKek.Models{
     public class ResultData{
         public Biodata biodata;
         public String algorithm;
-        public float Similarity;
+        public double Similarity;
         public float execTime;
         public string fingerprintImagePath;
 
@@ -15,7 +15,7 @@ namespace Tubes3_apaKek.Models{
         }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider adding the 'required' modifier or declaring as nullable.
 
-        public ResultData(Biodata biodata, string algorithm, float similarity, float execTime, string fingerprintImagePath){
+        public ResultData(Biodata biodata, string algorithm, double similarity, float execTime, string fingerprintImagePath){
             this.biodata = biodata;
             this.algorithm = algorithm;
             this.Similarity = similarity;
